fix(bashrc): use functions instead of aliases
For aliases, all the `$()` bits needed to be escaped so that the command
was run each time I used the alias. Otherwise, the command was run
initially when `bashrc` was sourced, and so my vit related aliases
didn't update their time periods correctly.

Anyway, for anything complex, functions should be preferred, so I'm just
going to use functions now instead of aliases.